Choose camera (#195)
* Choose a camera facing direction on widget creation

* Updated README with cameraFacing

* Renamed CameraFacing to CameraDirection

* Added directions to demo

* Added close image after read frame

* Changed to close latest frame on stop

* added frame.close to detector after analise

* added null check

* Added fix for orientation on android sdk 23 - 6.0

* Added fix for orientation on android sdk 23 - 6.0

* added null check

* Added fix for orientation on android sdk 23 - 6.0

* Fix android & implement for ios
