+++ android/src/main/java/com/github/rmtmckenzie/qrmobilevision/QrCamera.java
@@ -1,7 +1,7 @@
 package com.github.rmtmckenzie.qrmobilevision;
 
 interface QrCamera {
-    void start() throws QrReader.Exception;
+    void start(final int cameraDirection) throws QrReader.Exception;
     void stop();
     void toggleFlash();
     int getOrientation();
