+++ android/src/main/java/com/github/rmtmckenzie/qrmobilevision/QrMobileVisionPlugin.java
@@ -121,6 +121,7 @@ public class QrMobileVisionPlugin implements MethodCallHandler, QrReaderCallback
                     lastHeartbeatTimeout = methodCall.argument("heartbeatTimeout");
                     Integer targetWidth = methodCall.argument("targetWidth");
                     Integer targetHeight = methodCall.argument("targetHeight");
+                    Integer cameraDirection = methodCall.argument("cameraDirection");
                     List<String> formatStrings = methodCall.argument("formats");
 
                     if (targetWidth == null || targetHeight == null) {
@@ -137,7 +138,8 @@ public class QrMobileVisionPlugin implements MethodCallHandler, QrReaderCallback
                     readingInstance = new ReadingInstance(reader, textureEntry, result);
                     try {
                         reader.start(
-                            lastHeartbeatTimeout == null ? 0 : lastHeartbeatTimeout
+                            lastHeartbeatTimeout == null ? 0 : lastHeartbeatTimeout,
+                            cameraDirection == null ? 0 : cameraDirection
                         );
                     } catch (IOException e) {
                         e.printStackTrace();
