+++ lib/src/frame_rotation.dart
@@ -1,0 +1,1 @@
+enum FrameRotation { none, ninetyCC, oneeighty, twoseventyCC }
