+++ pubspec.yaml
@@ -20,3 +20,5 @@ dependencies:
   flutter:
     sdk: flutter
   native_device_orientation: ^1.0.0
+  device_info_plus: ^3.0.0
+
